tsx files deleted or renamed
