+++ components/LayOuts/Header/DropdownMenu.jsx
@@ -1,6 +1,4 @@
-
-import React, { useState } from 'react';
-
+import React, { useState } from "react";
 
 const DropdownMenu = ({ options, onSelect }) => {
   const [isOpen, setIsOpen] = useState(false);
